+++ api/model.js
@@ -1,4 +1,4 @@
-const OptimismModel = require('./data/optimismModel');
+const OptimismModel = require('./model/optimismModel');
 const knex = require('./db');
 
 const mainModel = new OptimismModel(knex);
