+++ api/model/slotItems.js
